% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_classes.R
\docType{data}
\name{age_classes}
\alias{age_classes}
\title{Key between ages and age classes}
\format{
A tibble
\describe{
 \item{age}{Age in years}
 \item{age_class}{Age class}
}
}
\usage{
age_classes
}
\description{
Key between ages and age classes
}
\keyword{datasets}
