% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution_ita.R
\docType{data}
\name{air_pollution_ita}
\alias{air_pollution_ita}
\title{background PM2.5 concentration levels in Italian cities}
\format{
A tibble
\describe{
 \item{city}{City name}
 \item{pm25}{Average PM2.5 concentration (ug/m3)}
}
}
\source{
\url{https://www.who.int/data/gho/data/themes/air-pollution/who-air-quality-database}
}
\usage{
air_pollution_ita
}
\description{
Data are average of the last ten years of available data from the WHO
}
\keyword{datasets}
