% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alive_pop.R
\name{alive_pop}
\alias{alive_pop}
\title{Compute the alive population given a starting population of a vector of mortality rates}
\usage{
alive_pop(starting_pop, death_rate)
}
\arguments{
\item{starting_pop}{A numeric value. The starting population.}

\item{death_rate}{A numeric vector. Annual mortality rates.}
}
\value{
A numeric vector
}
\description{
Compute the alive population given a starting population of a vector of mortality rates
}
