% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codici_comuni_2023.R
\docType{data}
\name{codici_comuni_2023}
\alias{codici_comuni_2023}
\title{Region, province and municipality statistical codes}
\format{
A tibble
\describe{
 \item{codice_regione}{Region code}
 \item{Codice dell'Unità territoriale sovracomunale (valida a fini statistici)}{Old regional code}
 \item{Codice Provincia (Storico)(1)}{Old province code}
 \item{Progressivo del Comune (2)}{Progressive municipality code}
 \item{codice_comune}{Municipality code}
 \item{Denominazione (Italiana e straniera)}{Municipality official name}
 \item{Denominazione in italiano}{Municipality name in Italian}
 \item{Denominazione altra lingua}{Municipality name in other language}
 \item{Codice Ripartizione Geografica}{Macro-area code}
 \item{Ripartizione geografica}{macro-area name}
 \item{Denominazione Regione}{Region name}
 \item{Provincia}{Province name}
 \item{Tipologia di Unità territoriale sovracomunale}{Type of territorial unit}
 \item{Flag Comune capoluogo di provincia/città metropolitana/libero consorzio}{name of province capital}
 \item{Sigla automobilistica}{Car plate code}
 \item{Codice Comune formato numerico}{Municipality code, alphanumeric}
 \item{Codice Comune numerico con 110 province (dal 2010 al 2016)}{Municipality code, 110 provinces}
 \item{Codice Comune numerico con 107 province (dal 2006 al 2009)}{Municipality code, 107 provinces}
 \item{Codice Comune numerico con 103 province (dal 1995 al 2005)}{Municipality code, 103 provinces}
 \item{Codice Catastale del comune}{Land registry code}
 \item{Codice NUTS1 2010}{NUTS 1 code, 2010}
 \item{Codice NUTS2 2010 (3)}{NUTS 2 code, 2010}
 \item{Codice NUTS3 2010}{NUTS 3 code, 2010}
 \item{Codice NUTS1 2021}{NUTS 1 code, 2021}
 \item{Codice NUTS2 2021 (3)}{NUTS 2 code, 2021}
 \item{Codice NUTS3 2021}{NUTS 3 code, 2021}
}
}
\source{
<https://esploradati.istat.it/databrowser/>
}
\usage{
codici_comuni_2023
}
\description{
Region, province and municipality statistical codes
}
\keyword{datasets}
