% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_yld.R
\name{compute_yld}
\alias{compute_yld}
\title{Compute the (discounted) amount of years lost to disability}
\usage{
compute_yld(acute_dw = 0.5, chronic_dw = 0.1, sdr = 0.02, disease_length = 25)
}
\arguments{
\item{acute_dw, chronic_dw}{Acute (first year) and chronic disability weights}

\item{sdr}{The social discount rate}

\item{disease_length}{Average disease length on average}
}
\value{
The discounted amount of years lost to disability
}
\description{
This function computes the years lost to disability (YLD) taking into account disease-specific
disability weights provided, a country-specific social discount rate and the average disease length.
Such parameters are provided by the user.
}
