% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_yll.R
\name{compute_yll}
\alias{compute_yll}
\title{Compute the discounted number of years of life lost due ti premature death}
\usage{
compute_yll(disease_length = 15, life_exp_red = 20, sdr = 0.02)
}
\arguments{
\item{disease_length}{Average disease length on average}

\item{life_exp_red}{Reduction in life expectancy due to disease (years)}

\item{sdr}{The social discount rate}
}
\value{
The discounted amount of years of life lost
}
\description{
This function computes the years of life lost (YLL) taking into disease length (to establish the
discounting factor), a user-defined social discount rate and the reduction in expected reduction
in life expectancy,
}
