% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demographic data on Italian province capitals}
\format{
A tibble
\describe{
 \item{nuts2}{NUTS2 code}
 \item{city}{Name of province capital}
 \item{age}{Age group}
 \item{life_exo}{Life expectancy at a given age (both sexes)}
 \item{deat_rate}{Death rate at a given age}
}
}
\source{
<https://esploradati.istat.it/databrowser/>
}
\usage{
demo_data
}
\description{
Demographic data on Italian province capitals
}
\keyword{datasets}
