% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disability_weights.R
\docType{data}
\name{disability_weights}
\alias{disability_weights}
\title{Disability weights of diseases and injuries considered}
\format{
A tibble
\describe{
 \item{disease}{Disease/Injury name}
 \item{phase}{Phase, acute or chronic}
 \item{disability_weight}{Disability weight (0-1)}
 \item{description}{Description of condition}
 \item{source}{Source}
}
}
\source{
\url{https://www.healthdata.org/research-analysis/gbd}
}
\usage{
disability_weights
}
\description{
Disability weights of diseases and injuries considered
}
\keyword{datasets}
