% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_incidence.R
\docType{data}
\name{disease_incidence}
\alias{disease_incidence}
\title{Disease incidence in Italy, latest data from Global Burden of Disease}
\format{
A tibble
\describe{
 \item{sex_name}{Sex}
 \item{cause_name}{Name of condition}
 \item{val}{Incidence, mean value}
 \item{upper}{Incidence, upper value}
 \item{lower}{DIncidence, lower value}
 \item{ages}{Age}
}
}
\source{
\url{https://www.healthdata.org/research-analysis/gbd}
}
\usage{
disease_incidence
}
\description{
Disease incidence in Italy, latest data from Global Burden of Disease
}
\keyword{datasets}
