% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_length.R
\docType{data}
\name{disease_length}
\alias{disease_length}
\title{Disease length}
\format{
A tibble
\describe{
 \item{disease}{Disease name}
 \item{min}{Lower bound of age group}
 \item{max}{Upper bound of age group}
 \item{lifelong}{is the disease lifelong? Yes or No}
 \item{duration_years}{Disease length (if not lifelong)}
 \item{life_exp_decrease}{Life expectancy decrease}
 \item{ages}{Age}
}
}
\source{
Various sources
}
\usage{
disease_length
}
\description{
Disease length
}
\keyword{datasets}
