% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_res_air.R
\name{dose_res_air}
\alias{dose_res_air}
\title{Dose-response functions - air pollution}
\usage{
dose_res_air(
  city = "Roma",
  mode_from = "car",
  mode_to = "bike",
  w_phy_act = c(5, 7, 8, 2, 1),
  w_commuting_from = seq(4, 20, by = 4),
  w_commuting_to = seq(6, 30, by = 6),
  experimental_data = FALSE,
  working_weeks = 48,
  commuting_days = 4,
  dimension = "morbidity"
)
}
\arguments{
\item{city}{A single character string. The name (in Italian) of the Italian municipality for which the
scenario must be built. THe full list of city names can be found in the commuting matrix
(see `comm_matrix_cities_km`)}

\item{mode_from, mode_to}{Transportation modes in the baseline and alternative scenarios (only needed for
the final output tibble, not involved in the computations)
and commuting (hours)}

\item{w_phy_act, w_commuting_from, w_commuting_to}{Weekly time spent performing leisure physical activity,
and commuting (both in the original and active mobility scenarios).}

\item{experimental_data}{Indicates whether experimental data or data from the literature should be used to
estimate the ventilation rate of `mode_to`.}

\item{working_weeks}{Number of annual working weeks}

\item{commuting_days}{Number of weekly commuting days. Default is `4`}

\item{dimension}{Indicates whether the dose-response function refers to non-communicable diseases (`morbidity`) related to
air pollution or all-cause mortality (`mortality`). Default value is `morbidity`.}
}
\description{
This function computes dose-response functions for air pollution. Such function
relates exposure to air pollution (PM2.5) to health outcomes (namely incidence of some
non-communicable diseases and all-cause mortality).
}
