% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_data.R
\docType{data}
\name{exp_data}
\alias{exp_data}
\title{Data from RSE experimentation on active mobility}
\format{
A tibble
\describe{
 \item{mode}{Transportation mode}
 \item{kcal}{PEnergy consumption (kcal)}
 \item{h}{Travel time (hours)}
 \item{user}{User name}
 \item{kg}{User weight}
 \item{bpm}{Uder average heart beat}
}
}
\usage{
exp_data
}
\description{
Data are collected by RSE workers commuting to work by bike or ebike
}
\keyword{datasets}
