% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatal_road_inj_rate.R
\docType{data}
\name{fatal_road_inj_rate}
\alias{fatal_road_inj_rate}
\title{Fatal road injury rate}
\format{
A tibble
\describe{
 \item{city}{City name}
 \item{province}{Province name}
 \item{mode}{Transportation mode}
 \item{age_class}{Age class}
 \item{region}{Region name)}
 \item{geo}{Macro-area}
 \item{fatal_road_inj_rate}{Rate of fatal road injuries}
}
}
\source{
Own calculation on ISTAT data
}
\usage{
fatal_road_inj_rate
}
\description{
Fatal road injury rate
}
\keyword{datasets}
