% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_lit.R
\docType{data}
\name{met_lit}
\alias{met_lit}
\title{Metabolic equivalents (measure of physical effort)}
\format{
A tibble
\describe{
 \item{mode}{Transportation mode}
 \item{met}{metabolic equivalents (MET/h)}
}
}
\usage{
met_lit
}
\description{
Data from the literature
}
\keyword{datasets}
