% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_change_share.R
\name{mode_change_share}
\alias{mode_change_share}
\title{Return the share of commuters shifting from passive to active mobility}
\usage{
mode_change_share(
  km_one_way,
  share.less_3km = 0.3,
  share.3_10km = 0.2,
  share.10_15km = 0,
  share.more_15km = 0
)
}
\arguments{
\item{km_one_way}{A numeric value. The distance (in km) commuters have to travel every day (one way only)}

\item{share.less_3km}{A numeric value. The share of commuters commuting less than 3 km (one way)
making the modal switch.}

\item{share.3_10km}{A numeric value. The share of commuters commuting between 3 and 10 km (one way)
making the modal switch.}

\item{share.10_15km}{A numeric value. The share of commuters commuting between 10 and 15 km (one way)
making the modal switch.}

\item{share.more_15km}{A numeric value. The share of commuters commuting more than 15 km (one way)
making the modal switch.}
}
\value{
A numeric value.
}
\description{
Return the share of commuters shifting from passive to active mobility
}
