% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morbidity_impact.R
\name{morbidity_impact}
\alias{morbidity_impact}
\title{Compute the overall health impact of active mobility on both mortality and morbidity}
\usage{
morbidity_impact(
  city = "Roma",
  min_age = 20,
  max_age = 64,
  mode_from = "car",
  mode_to = "bike",
  share.less_3km = 0.3,
  share.3_10km = 0.2,
  share.10_15km = 0,
  share.more_15km = 0,
  commuting_days = 4,
  experimental_data = FALSE,
  met_phy_act = 7,
  working_weeks = 48,
  detail = "low",
  sdr = 0.024,
  ref_year = 2021,
  voly = 70151,
  vsl = 3607757
)
}
\arguments{
\item{city}{A single character string. The name (in Italian) of the Italian muncipality for which the
scenario must be built. THe full list of city names can be found in the commuting matrix
(see `data_raw$comm_matrix_cities_km`)}

\item{min_age, max_age}{Numeric values. The minimum and the maximum ages commuters can have to be included in the scenario.}

\item{mode_from, mode_to}{Character strings. The transportation mode from which commuters shift from. And
the transportation mode to which they shift to. Argument `mode_to` can assume the following values:
`walk`, `bike` and `ebike`. Argument `mode_from` can assume any value amongst those included in
the commuting matrix: `r knitr::combine_words(setdiff(data_raw$transport_speeds$mean_of_transp, c("walk", "bike", "ebike")))`}

\item{share.less_3km}{A numeric value. The share of commuters commuting less than 3 km (one way)
making the modal switch.}

\item{share.3_10km}{A numeric value. The share of commuters commuting between 3 and 10 km (one way)
making the modal switch.}

\item{share.10_15km}{A numeric value. The share of commuters commuting between 10 and 15 km (one way)
making the modal switch.}

\item{share.more_15km}{A numeric value. The share of commuters commuting more than 15 km (one way)
making the modal switch.}

\item{commuting_days}{A numeric value. The number of days in which new adopters perform travels to and
from the place of work/study.}

\item{experimental_data}{A logical value. If set to `FALSE`, the default, MET values are taken
from the literature. If set to `TRUE` values from the RSE experimentation are used.}

\item{met_phy_act}{A numeric value. The physical effort (measured in MET/h) associated with physical
activity performed by individuals in their free time (default is set to `7`).}

\item{working_weeks}{Number of annual working weeks. Default is `48`}

\item{detail}{If set to `low`, the default, the function returns results the overall health impact expressed
in euro/km. Other options are: `medium` and `high`, which return results with detail at increasing level
of detail: by dimension (physical activity, air pollution and road injuries) or by disease/outcome respectively.}

\item{sdr}{The social discount rate to be used to discount future YLL and YLD. The default is `0.024`.
The function `sdr` of the `btransfer` package can be used to compute the social discount rate}

\item{ref_year}{Year of reference of the analysis. Default is `2021`}

\item{voly, vsl}{The value of the VOLY or the VSL to be used. Values adjusted to a user-defined policy year can
be computed with function `actranspr2::voly_vsl_calculator`}
}
\value{
A tibble
}
\description{
This function calculates the impact of a shift from passive to active mobility on morbidity and
mortality for Italian municipalities taking into account the effect of increased physical activity,
increased air pollution intake and increased road injury risk.
}
\examples{
morbidity_impact(
mode_from = "car",
mode_to = "ebike",
detail = "medium"
)
}
