% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morbidity_rr.R
\docType{data}
\name{morbidity_rr}
\alias{morbidity_rr}
\title{Dose-response functions linking physical activity and morbidity}
\format{
A tibble
\describe{
 \item{outcome}{Outcome}
 \item{sex}{Sex}
 \item{dimension}{Physical activity or air pollution}
 \item{rr}{Relative risk corresponding to exposure}
 \item{exposure}{Exposure}
 \item{um_unit}{Exposure measurement unit}
 \item{trans_of_exp}{Non-linear transformation of exposure}
 \item{source}{Source}
}
}
\usage{
morbidity_rr
}
\description{
Dose-response functions linking physical activity and morbidity
}
\keyword{datasets}
