% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rr.R
\docType{data}
\name{mortality_rr}
\alias{mortality_rr}
\title{Dose-response functions linking physical activity and mortality}
\format{
A tibble
\describe{
 \item{outcome}{Outcome}
 \item{mode}{Transportation mode}
 \item{dimension}{Physical activity or air pollution}
 \item{rr}{Relative risk corresponding to exposure}
 \item{exposure}{Exposure}
 \item{um_unit}{Exposure measurement unit}
 \item{trans_of_exp}{Non-linear transformation of exposure}
 \item{threshold}{Threshold of exposure above which there are no addition effects}
 \item{source}{Source}
}
}
\usage{
mortality_rr
}
\description{
Dose-response functions linking physical activity and mortality
}
\keyword{datasets}
