% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_act.R
\docType{data}
\name{phy_act}
\alias{phy_act}
\title{Physical activity levels by age group}
\format{
A tibble
\describe{
 \item{age}{NAge group}
 \item{phy_act_weekly_h}{Weekly hours of leisure physical activity}
 \item{share_phy_act}{Share of total population}
}
}
\source{
\url{https://ec.europa.eu/eurostat/data/database}
}
\usage{
phy_act
}
\description{
Number of weekly hours of leisure physical activity
}
\keyword{datasets}
