% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm25_concentration.R
\name{pm25_concentration}
\alias{pm25_concentration}
\title{Provide the average PM2.5 concentration for Italian municipalities}
\usage{
pm25_concentration(place = "Roma")
}
\arguments{
\item{place}{A string character. The name (in Italian) of provice capital of interest.}
}
\value{
A numeric value: the average PM2.5 background concentration expressed in um/m3.
}
\description{
This function returns the average (2010-2019) PM2.5 background concentration for Italian province capitals.
Data are from the WHO Air Pollution Database (2022 version). If data at city level are not available,
the function returns the average value of all cities in the same region.
}
