% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_density.R
\docType{data}
\name{pop_density}
\alias{pop_density}
\title{Population density of Italian province capitals}
\format{
A tibble
\describe{
 \item{city}{City name}
 \item{km2}{Municipality area}
 \item{pop}{Total population}
 \item{pop_density}{Population density}
}
}
\source{
\url{www.istat.it}
}
\usage{
pop_density
}
\description{
Data are from ISTAT
}
\keyword{datasets}
