% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_capitals.R
\docType{data}
\name{regional_capitals}
\alias{regional_capitals}
\title{Regional and province capitals}
\format{
A tibble
\describe{
 \item{city}{City name}
 \item{prov_capital}{Province capital name}
 \item{region_capital}{Regional capital name}
 \item{province}{Province name}
 \item{region}{Region name}
 \item{geo}{Macro-area}
}
}
\usage{
regional_capitals
}
\description{
Regional and province capitals
}
\keyword{datasets}
