% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_inj_rate.R
\docType{data}
\name{road_inj_rate}
\alias{road_inj_rate}
\title{Road injury rates (non-fatal injuries)}
\format{
A tibble
\describe{
 \item{city}{Name of province capital}
 \item{province}{Province name}
 \item{mode}{Transportation mode}
 \item{age_class}{Age group}
 \item{region}{Region name}
 \item{geo}{Macro-region name}
 \item{outcome}{Injury description}
 \item{road_inj_rate}{Road injury rate, expressed in accidents/km}
}
}
\usage{
road_inj_rate
}
\description{
Road injury rates (non-fatal injuries)
}
\keyword{datasets}
