% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_injuries.R
\docType{data}
\name{road_injuries}
\alias{road_injuries}
\title{Road_injuries in Italian province capitals}
\format{
A tibble
\describe{
 \item{city}{City name}
 \item{prov}{Province}
 \item{mode}{Transportation mode}
 \item{age}{Agw group}
 \item{pop}{Population}
 \item{injury}{injuries (average of the last 5 years)}
 \item{death}{deaths (average of the last 5 years)}
}
}
\source{
\url{www.istat.it}
}
\usage{
road_injuries
}
\description{
Data are from ISTAT
}
\keyword{datasets}
