% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_volumes.R
\docType{data}
\name{traffic_volumes}
\alias{traffic_volumes}
\title{Traffic_volumes in Italian province capitals}
\format{
A tibble
\describe{
 \item{city}{City name}
 \item{mode}{Transportation mode}
 \item{traffic_volume}{Traffic volumes}
 \item{volume_share}{Share of total traffic}
}
}
\source{
\url{www.istat.it}
}
\usage{
traffic_volumes
}
\description{
Data are from ISTAT (commuting matrix 2011)
}
\keyword{datasets}
