% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent_rates.R
\docType{data}
\name{vent_rates}
\alias{vent_rates}
\title{Average ventilation rates of different activities}
\format{
A tibble
\describe{
 \item{activity}{Activity description}
 \item{vent_rates}{Ventilation rate (m3/h)}
 \item{conf_fct}{Adjustment factors}
}
}
\usage{
vent_rates
}
\description{
Data include commuting modes and adjustment factors to adjust background concentrations depending on the
environment in which an activity takes place
}
\keyword{datasets}
